requirements.txt
asgiref==3.4.1
Django==3.2.9
numpy==1.21.4
opencv-python==4.5.4.58
Pillow==8.4.0
pytesseract==0.3.8
pytz==2021.3
sqlparse==0.4.2
